requirements.txt
numpy
sqlite-vec # vectorize columns , simple and easy , faiss could be used in mem
scikit-learn # for tdifd stuff
pandas # better data handling
flatten-json #helper cleaner code 
fastembed # embeddings version
llama-cpp-python # for the local model 

